android projects

+++ JavaTestMahesh/src/com/mahesh/exceptionHandelling/ClassCast.java
@@ -3,7 +3,7 @@ package com.mahesh.exceptionHandelling;
 public class ClassCast {
 	public static void main(String[] args) {
 		try {
-			Object obj = new Integer(100);
+			Object obj = new Integer(10);
 			System.out.println((float) obj);
 		} catch (ClassCastException e) {
 			System.out.println("Check the casting");
